pyproject.toml
[tool.poetry]
name = "test-da"
version = "0.1.0"
description = ""
authors = ["Nicolai94 <nicolas9441@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.108.0"}
sqlalchemy = "^2.0.24"
alembic = "^1.13.1"
pytest = "^7.4.4"
structlog = "^23.3.0"
pyhumps = "^3.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
